Image_models/Density_models/S-DCNet/requirements.txt
python==3.6.2
pytorch>=0.4.0
numpy==1.14.0
scikit-image==0.13.1
scipy==1.0.0
pandas==0.22.0